my_flask_api_full/requirements.txt
# Web
Flask~=3.0
gunicorn~=22.0

# Queue / Cache
pika~=1.3
redis~=5.0

# Elasticsearch client (8.x cluster)
elasticsearch~=8.14

# Auth & crypto
PyJWT~=2.8
jwcrypto~=1.5

# HTTP / utils
requests~=2.32
jsonschema~=4.22
python-dotenv~=1.0
